backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.3
greenlet==3.3.1
h11==0.16.0
idna==3.11
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-multipart==0.0.22
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles>=0.21.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
email-validator>=2.0.0
slowapi>=0.1.9
